moved Calendar.js to components
